Refactoring changes post removal of customers Hashmap

+++ server/Account.java
@@ -25,7 +25,7 @@ public class Account {
         return accountType;
     }
 
-    public double getBalance() {
+    public int getBalance() {
         return balance;
     }
